requirements.txt
streamlit==1.32.0
pandas==2.1.4
plotly==5.18.0
fpdf2==2.7.7
python-dotenv==1.0.0
supabase==2.3.1